// src/app/models/assessment-configuration.ts
export class AssessmentConfiguration {
    id: number;
    domainSelector: string;
    domain: string;
    subject: string;
    question: string;
    metric: string;
    type: string;
    tin: string;
    bronze: string;
    silver: string;
    gold: string;
    platinum: string;
    status: string;
}